Updated log prefix syntax
Shouldn't affect many scripts, as they rely on an SQL db of paths which
won't change. This means that scripts will be stored in path in time
order (thanks to the zero padding) and won't have extraneous information
contained in the name (control scheme is too compleex for the simple
control name originally given).